First 50 fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,17 +7,17 @@
  */
 int main(void)
 {
-	int a = 1, b = 2;
+	unsigned long int a = 1, b = 2;
 	unsigned long int f = 0;
 	int i = 2;
 
-	printf("%d, %d", a, b);
+	printf("%lu, %lu", a, b);
 	while (i < 50)
 	{
 		f = a + b;
 		a = b;
 		b = f;
-		printf(", %ld", f);
+		printf(", %lu", f);
 		i++;
 	}
 	putchar('\n');
